added lat and lng to survey

+++ database/migrations/2016_02_20_000951_create_surveys_table.php
@@ -31,8 +31,8 @@ class CreateSurveysTable extends Migration
             $table->text('facultytimeallotted');
             $table->text('adminsupport');
             $table->text('contactinfo');
-            $table->float('lat');
-            $table->float('lng');
+            $table->double('lat');
+            $table->double('lng');
 
 
             $table->timestamps();
